+++ core/block/import/converter/filenameprovider.go
@@ -7,7 +7,7 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/anyproto/anytype-heart/core/block/import/source"
+	"github.com/anyproto/anytype-heart/core/block/import/converter/source"
 	"github.com/anyproto/anytype-heart/pkg/lib/core"
 	oserror "github.com/anyproto/anytype-heart/util/os"
 )
