+++ core/block/import/converter/syncer/icon.go
